[#1159] Move helper functions to greedybear/utils and update imports
