+++ greedybear/cronjobs/repositories/sensor.py
@@ -30,7 +30,7 @@ class SensorRepository:
         Returns:
             Sensor object if valid, None if invalid IP format.
         """
-        from greedybear.cronjobs.extraction.utils import get_ioc_type
+        from greedybear.utils import get_ioc_type
 
         if ip in self.cache:
             return self.cache[ip]
